+++ backend/app/controllers/student_controller.py
@@ -1,6 +1,5 @@
 from pymongo import MongoClient
 from bson.objectid import ObjectId
-from flask import current_app
 from dotenv import load_dotenv, find_dotenv
 import os
 
@@ -11,6 +10,7 @@ def get_student_by_id(student_id):
     client = MongoClient(os.getenv("MONGO_URI"))
     db = client["vandytime_db"]
     student_collection = db["students"]
+
     try:
         student_object = student_collection.find_one({"_id": ObjectId(student_id)})
         return student_object
@@ -26,7 +26,11 @@ def create_student(student_data):
     client = MongoClient(os.getenv("MONGO_URI"))
     db = client["vandytime_db"]
     student_collection = db["students"]
+
     try:
+        if "grades" not in student_data:
+            student_data["grades"] = []
+
         student_object = student_collection.insert_one(student_data)
         return student_object
     except Exception as e:
@@ -41,6 +45,7 @@ def update_student(student_id, update_data):
     client = MongoClient(os.getenv("MONGO_URI"))
     db = client["vandytime_db"]
     student_collection = db["students"]
+
     try:
         student_object = student_collection.update_one({"_id": ObjectId(student_id)}, {"$set": update_data})
         return student_object
@@ -56,15 +61,21 @@ def delete_student(student_id):
     client = MongoClient(os.getenv("MONGO_URI"))
     db = client["vandytime_db"]
     student_collection = db["students"]
+
     try:
         student_data = get_student_by_id(student_id)
+
+        if not student_data:
+            return {"deleted_count": 0}
+
         student_grades = student_data["grades"]
         for grade_id in student_grades:
             delete_grade(grade_id)
+            
         delete_object = student_collection.delete_one({"_id": ObjectId(student_id)})
         return {"deleted_count": delete_object.deleted_count}
     except Exception as e:
         print(f"Error deleting student: {e}")
-        return None
+        return {"error": str(e)}
     finally:
         client.close()
